+++ src/main/java/com/example/EcoRadar/repository/GreenAreaAddressRepository.java
@@ -4,5 +4,5 @@ import com.example.EcoRadar.model.entity.GreenAreaAddress;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GreenAreaAddressRepository
-        extends JpaRepository<GreenAreaAddress, Long> {
+        extends JpaRepository<GreenAreaAddress, Integer> {
 }
