+++ src/main/java/com/example/EcoRadar/repository/GreenAreaHistoryRepository.java
@@ -4,5 +4,5 @@ import com.example.EcoRadar.model.entity.GreenAreaHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GreenAreaHistoryRepository
-        extends JpaRepository<GreenAreaHistory, Long> {
+        extends JpaRepository<GreenAreaHistory, Integer> {
 }
