+++ src/main/java/com/example/EcoRadar/repository/GreenAreaRepository.java
@@ -4,5 +4,5 @@ import com.example.EcoRadar.model.entity.GreenArea;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GreenAreaRepository
-        extends JpaRepository<GreenArea, Long> {
+        extends JpaRepository<GreenArea, Integer> {
 }
