+++ src/main/java/com/example/EcoRadar/repository/UserAddressRepository.java
@@ -4,5 +4,5 @@ import com.example.EcoRadar.model.entity.UserAddress;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserAddressRepository
-        extends JpaRepository<UserAddress, Long> {
+        extends JpaRepository<UserAddress, Integer> {
 }
