+++ src/main/java/com/example/EcoRadar/repository/UserRepository.java
@@ -4,5 +4,5 @@ import com.example.EcoRadar.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository
-        extends JpaRepository<User, Long> {
+        extends JpaRepository<User, Integer> {
 }
